requirements.txt
streamlit
google-generativeai
music21
midiutil
